+++ core/src/main/java/io/github/stardewmini/common/model/GameData.java
@@ -1,9 +1,10 @@
 package io.github.stardewmini.common.model;
 
 import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 
-public record GameData(String[] playerNames, int[] playerFarms, int seed){
+public record GameData(String[] playerNames, int[] playerFarms, long seed){
     @Override
     public boolean equals(Object o) {
         if (!(o instanceof GameData gameData)) return false;
